refactor(skippy): isolate binary connection session tracking
Keep orphaned-lane reclamation separate from the already oversized binary connection message loop, with its focused unit tests beside the tracker.
